week_4/requirements.txt
requests==2.26.0
google-cloud-storage==1.42.0
